pyproject.toml
[project]
name = "parking-management-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "boto3>=1.42.18",
    "boto3-stubs~=1.42.18",
    "botocore[crt]>=1.42.18",
    "fastapi[standard]>=0.128.0",
    "mypy==1.19.1",
    "mypy-boto3-dynamodb>=1.42.3",
    "pyjwt>=2.10.1",
    "pyrefly>=0.47.0",
    "pyright==1.1.407",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
    "types-boto3>=1.42.18",
]
